ci: fix automated NPM release
