+++ .github/workflows/release.yml
@@ -2,14 +2,14 @@ name: Release
 
 on:
   push:
-    tags:
-      - 'v*'
+    branches: [main]
 
 permissions:
   contents: write
 
 jobs:
   release:
+    if: "contains(github.event.head_commit.message, 'release: v')"
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
@@ -24,6 +24,16 @@ jobs:
         with:
           node-version-file: .node-version
           cache: pnpm
+
+      - name: Extract version from commit message
+        run: |
+          VERSION=$(echo "${{ github.event.head_commit.message }}" | grep -oP 'release: \Kv[0-9]+\.[0-9]+\.[0-9]+')
+          echo "VERSION=$VERSION" >> $GITHUB_ENV
+
+      - name: Create and push tag
+        run: |
+          git tag ${{ env.VERSION }}
+          git push origin ${{ env.VERSION }}
 
       - run: npx changelogithub
         continue-on-error: true
