+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "eslint-plugin-oxlint",
-  "version": "0.5.0",
+  "version": "0.6.1",
   "description": "Turn off all rules already supported by oxlint",
   "type": "module",
   "packageManager": "pnpm@9.6.0",
@@ -36,7 +36,7 @@
   },
   "license": "MIT",
   "scripts": {
-    "generate": "node --import @oxc-node/core/register ./scripts/generate.ts",
+    "generate": "node --import @oxc-node/core/register ./scripts/generate.ts && pnpm format",
     "clone": "node --import @oxc-node/core/register ./scripts/sparse-clone.ts",
     "build": "vite build",
     "lint": "npx oxlint && npx eslint",
@@ -68,7 +68,7 @@
     "husky": "^9.0.11",
     "lint-staged": "^15.2.5",
     "memfs": "^4.9.2",
-    "oxlint": "^0.7.0",
+    "oxlint": "^0.6.1",
     "prettier": "^3.3.1",
     "shelljs": "^0.8.5",
     "typescript": "^5.4.5",
